Extend options object with defaults
Using defaults as a function parameter causes the whole options object to
be overriden, even when the intent is to change just one parameter.

Using the spread operator allows a overriding only individual options.

An alternative could be using destructing with defaults as the function argument
like this: `(id, url, { force = false, method = 'GET' } = {})`, but that
requires to also set default values for the `extendAction()` function
(`headers` and `credentials`).

Resolves #14